refactor(ui): merge Settings + Profile into one Account page
Neither page earned a nav slot after the honest pruning: Settings was a
theme picker + MFA card, and Profile still shipped dead Change
Password / Configure MFA buttons plus a fake edit mode whose Save
persisted nothing (no backend endpoints exist for any of it).

Account = identity/auth status + appearance + real MFA setup. Routes:
/account replaces /settings; /settings, /profile, and /config redirect
there. User-menu item updated.
